context_propagation/python_gateway/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai>=1.12.0
requests>=2.31.0
arize-otel==0.5.3
openinference-instrumentation-openai==0.1.17
opentelemetry-sdk==1.28.1
opentelemetry-exporter-otlp==1.28.0
opentelemetry-instrumentation==0.49b0
opentelemetry-api>=1.28.0
opentelemetry-semantic-conventions>=0.49b0
openinference-semantic-conventions>=0.1.5
opentelemetry-proto>=1.28.0
protobuf>=5.28.3
googleapis-common-protos~=1.52
grpcio>=1.63.2
arize-phoenix-otel>=0.1.0 